refactor: MemberService, MemberServiceImpl 코드 정리
