+++ src/test/java/kr/co/swm/member/model/service/MemberServiceTest.java
@@ -23,7 +23,7 @@ class MemberServiceTest {
     private SmsCertificationUtil smsCertificationUtil;
 
     @InjectMocks
-    private MemberService memberService;
+    private MemberServiceImpl memberService;
 
     @BeforeEach
     void setUp() {
